# backend/src/routes/public.routes.js
import { Router } from 'express';
import rateLimit from 'express-rate-limit';
import { z } from 'zod';
import { env } from '../config/env.js';
import { validate } from '../middlewares/validate.middleware.js';
import * as articles from '../services/article.service.js';
import { asyncHandler,success } from '../utils/http.js';
const router=Router();const slug=z.object({slug:z.string().min(1).max(300)});const page=z.object({page:z.coerce.number().int().positive().optional(),limit:z.coerce.number().int().positive().max(50).optional()});
router.get('/homepage',asyncHandler(async(_req,res)=>{res.set('Cache-Control','public, max-age=30, s-maxage=60, stale-while-revalidate=60');return success(res,await articles.homepage());}));
router.get('/categories/:slug',validate({params:slug}),asyncHandler(async(req,res)=>success(res,await articles.getCategory(req.params.slug))));
router.get('/categories/:slug/articles',validate({params:slug,query:page}),asyncHandler(async(req,res)=>{const data=await articles.categoryArticles(req.params.slug,req.query);return success(res,data.items,{meta:{...data.meta,category:data.category}});}));
router.get('/articles/trending',validate({query:z.object({period:z.enum(['24h','7d']).default('24h'),limit:z.coerce.number().int().positive().max(50).default(20)})}),asyncHandler(async(req,res)=>success(res,await articles.getTrending(req.query.period,req.query.limit))));
router.get('/articles/:slug/views',validate({params:slug}),asyncHandler(async(req,res)=>success(res,{views:await articles.getViews(req.params.slug)})));
router.post('/articles/:slug/views',rateLimit({windowMs:60_000,limit:60,standardHeaders:'draft-8',legacyHeaders:false}),validate({params:slug}),asyncHandler(async(req,res)=>success(res,{views:await articles.recordView(req.params.slug,req)})));
router.get('/articles/:slug',validate({params:slug}),asyncHandler(async(req,res)=>{const data=await articles.articleDetail(req.params.slug);if(data.redirectSlug)return res.redirect(301,`${env.WEB_URL}/tin-tuc/${data.redirectSlug}`);res.set('Cache-Control','public, max-age=60, s-maxage=300, stale-while-revalidate=300');return success(res,data);}));
router.get('/tags/:slug/articles',validate({params:slug,query:page}),asyncHandler(async(req,res)=>{const data=await articles.tagArticles(req.params.slug,req.query);return success(res,data.items,{meta:{...data.meta,tag:data.entity}});}));
router.get('/authors/:slug',validate({params:slug}),asyncHandler(async(req,res)=>{const data=await articles.authorArticles(req.params.slug,{page:1,limit:1});return success(res,data.entity);}));
router.get('/authors/:slug/articles',validate({params:slug,query:page}),asyncHandler(async(req,res)=>{const data=await articles.authorArticles(req.params.slug,req.query);return success(res,data.items,{meta:{...data.meta,author:data.entity}});}));
router.get('/search',validate({query:z.object({q:z.string().max(200).default(''),page:z.coerce.number().int().positive().optional(),limit:z.coerce.number().int().positive().max(50).optional()})}),asyncHandler(async(req,res)=>{const data=await articles.search(req.query);return success(res,data.items,{meta:data.meta});}));
router.get('/preview/:token',validate({params:z.object({token:z.string().min(20).max(2000)})}),asyncHandler(async(req,res)=>{res.set('X-Robots-Tag','noindex, nofollow');res.set('Cache-Control','private, no-store');return success(res,await articles.preview(req.params.token));}));
router.get('/rss',asyncHandler(async(_req,res)=>{const data=await articles.latestArticles(50);const items=data.map(item=>`<item><title><![CDATA[${item.title.replaceAll(']]>',']]]]><![CDATA[>')}]]></title><link>${env.WEB_URL}/tin-tuc/${item.slug}</link><pubDate>${new Date(item.publishedAt).toUTCString()}</pubDate></item>`).join('');res.type('application/rss+xml').send(`<?xml version="1.0" encoding="UTF-8"?><rss version="2.0"><channel><title>News Portal</title><link>${env.WEB_URL}</link>${items}</channel></rss>`);}));
export default router;